+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is roughly based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.19.0]
+
+- Support regexes in `collection_whitelist` config (eu-cdse/openeo-cdse-infra#54)
+
 ## [0.18.4]
 
-- Move `collection_whitelist` config to `AggregatorBackendConfig`  ([#112](https://github.com/Open-EO/openeo-aggregator/issues/112))
+- Move `collection_whitelist` config to `AggregatorBackendConfig` ([#112](https://github.com/Open-EO/openeo-aggregator/issues/112))
 
 ## [0.18.3]